added login input
Added a new `InputEmail` composable function for handling email input fields.
This new component is used in the `LoginScreen` to replace the generic `OutlinedTextField` for the email field.
Additionally, updated the `.gitignore` file to include `kotlin-js-store/wasm/yarn.lock` and added a comment to `build.gradle.kts` for running wasmJsBrowserDevelopment.