Develop (#4)
* Added kotlinx.serialization and setup server

This commit introduces the following changes:
- Added kotlinx.serialization as a dependency in `gradle/libs.versions.toml`.
- Configured the Ktor server in `server/src/main/kotlin/com/eugenerogov/planmind/Application.kt` to use JSON content negotiation with kotlinx.serialization.
- Added a new endpoint `/plan-mind` to the server which returns a list of names as JSON.
- Added a new `UseCaseModule.kt` in `server/src/main/kotlin/com/eugenerogov/planmind/di/` for Koin dependency injection, specifically for a `SignInLoginUseCase`.
- Updated dependencies in `server/build.gradle.kts` to include Koin, Arrow Core, coroutines, and serialization.
- Moved `projects.shared` dependency from `commonMain` to the top-level dependencies in `composeApp/build.gradle.kts`.

* added login input

Added a new `InputEmail` composable function for handling email input fields.
This new component is used in the `LoginScreen` to replace the generic `OutlinedTextField` for the email field.
Additionally, updated the `.gitignore` file to include `kotlin-js-store/wasm/yarn.lock` and added a comment to `build.gradle.kts` for running wasmJsBrowserDevelopment.

* Move compose deps to shared module

- Moves the composeMultiplatform and composeCompiler plugins from the platform-specific `build.gradle.kts` files to the shared `build.gradle.kts`.
- Adds compose-resources dependency.
- Updates the version alias for composeMultiplatform to compose-multiplatform.

* Remove Gson dependency

The Gson library has been removed from the project as it is no longer needed.

* Upgrade compose dependencies

This commit upgrades the following Compose dependencies:
- compose-multiplatform from 1.8.2 to 1.8.3
- compose from 1.8.2 to 1.8.3

It also updates the `compose-resources` and `compose-runtime` dependencies to use the new versions.

Additionally, this commit:
- Adds a `strings.xml` file with an `app_name` and `email_hint` string.
- Updates the `LoginScreen` to use the `email_hint` string resource.

* Move compose.resources dependency to shared module

The `compose.resources` dependency has been moved from the `composeApp/build.gradle.kts` to the `shared/build.gradle.kts`. This change makes the dependency available to all modules that depend on the shared module.

* Update libraries

- Decompose to 3.3.0
- Ktor from client to regular

---------

Co-authored-by: Eugene Rogov <eugeniirogov@MacBook-Pro-Eugenii.local>
Co-authored-by: Eugene Rogov <eugeniirogov@Mac.lan>