Develop (#9)
* Refactor: Update server endpoints

Moved server port constant to `Endpoint` object and updated endpoint paths for user profile and plan-mind.

* Refactor: Implement `TokenHolder` and `AuthPlugin` for token-based authentication

This commit introduces a new `TokenHolder` class to manage the authentication token. It also adds an `AuthPlugin` for Ktor, which automatically includes the token in the "Authorization" header of outgoing requests.

The `TokenHolder` and `AuthPlugin` are integrated into the DI setup.

Additionally, the `networkModule` has been renamed to `serviceApiModule` for better clarity.

* Added authentication mechanism using JWT

This commit introduces JWT-based authentication to the server.

Key changes:
- Added dependencies for Ktor server auth, JWT, and Auth0 Java JWT.
- Implemented a `/user/sign-in` endpoint that accepts login credentials (email and password).
- Upon successful login (hardcoded credentials for now), a JWT is generated and returned to the client.
- The JWT is configured with an audience, issuer, a "userId" claim, and an expiration time of 10 minutes.
- Configured JWT authentication in Ktor, including a verifier and validation logic.
- Created `LoginRequest.kt` data class for handling login requests.

* Refactor: move core remote classes to core package

Moved `TokenHolder`, `Endpoint`, and `AuthPlugin` to the `com.eugenerogov.planmind.data.remote.core` package.
Created `HostHolder` in the core package.

Introduced `AuthService` and `ProfileService` interfaces and their implementations.
Updated `ServiceApiModule` to reflect these changes and provide instances of the new services.

* feat: Move LoginRequest and add RegisterRequest

- Moved `LoginRequest.kt` from `domain/entities/profile` to `domain`.
- Added `RegisterRequest.kt` in `domain`.
- Updated import for `LoginRequest` in `Application.kt`.

* Implemented Koin module for ProfileViewModel

The ViewModelModule.kt file provides a Koin module for instantiating ProfileViewModel. It allows for injecting ComponentContext and an optional callback for when the profile is saved.

* Koin refactor: Extract Koin initialization into a common module

This commit refactors the Koin dependency injection setup by:

- Moving the Koin initialization logic into a new `KoinInit.kt` file within the `commonMain` source set.
- Calling the `initKoin()` function from the common `PlanMindApp` composable.
- Removing the Android-specific Koin initialization from the `App` class in `androidMain`.

This change centralizes Koin initialization, making it easier to manage and ensuring consistency across platforms.

---------

Co-authored-by: Eugene Rogov <eugeniirogov@MacBook-Pro-Eugenii.local>